backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.34.0"
fastapi = "^0.115.7"
langchain = "^0.3.16"
langchain-groq = "^0.2.3"
langchain-community = "^0.3.16"
langchain-core = "^0.3.32"
faiss-cpu = "^1.9.0.post1"
fastapi-cors = "^0.0.6"
python-dotenv = "^1.0.1"
langchain-huggingface = "^0.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
